Added: funcion

+++ src/model/observer.java
@@ -3,6 +3,6 @@ package model;
 import java.util.List;
 
 public interface observer {
-	public void onFinished(double[][] best, double[][] bestGen, double[][] average, List<Double> bestVars);
+	public void onFinished(double[][] best, double[][] bestGen, double[][] average, List<Integer> bestVars);
 	public void onNextGeneration();
 }
